Refactored the whole code Added support for private group like https://t.me/(joinchat/|\+)[invite_code]
