+++ telepathy/telepathy.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python3
 
+
 """Telepathy cli interface:
     An OSINT toolkit for investigating Telegram chats.
 """
@@ -18,6 +19,18 @@ import re
 import textwrap
 import time
 
+from libs.utils import (
+    print_banner,
+    color_print_green,
+    populate_user,
+    process_message,
+    parse_tg_date,
+    parse_html_page
+)
+import libs.const as const
+
+from colorama import Fore, Back, Style
+
 from telethon.errors import SessionPasswordNeededError, ChannelPrivateError
 from telethon.tl.types import (
     InputPeerEmpty,
@@ -31,131 +44,8 @@ from telethon.tl.types import (
 from telethon.tl.functions.messages import GetDialogsRequest
 from telethon import TelegramClient, functions, types, utils
 from telethon.utils import get_display_name, get_message_id
-from googletrans import Translator, constants
-from colorama import Fore, Back, Style
 from alive_progress import alive_bar
 from bs4 import BeautifulSoup
-from pprint import pprint
-
-__author__ = "Jordan Wildon (@jordanwildon)"
-__license__ = "MIT License"
-__version__ = "2.1.10"
-__maintainer__ = "Jordan Wildon"
-__email__ = "j.wildon@pm.me"
-__status__ = "Development"
-
-user_agent = [
-    # Chrome
-    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.113 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.90 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 5.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.90 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 6.2; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.90 Safari/537.36",
-    "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/44.0.2403.157 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 6.3; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.113 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/57.0.2987.133 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/57.0.2987.133 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.87 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.87 Safari/537.36",
-    # Firefox
-    "Mozilla/4.0 (compatible; MSIE 9.0; Windows NT 6.1)",
-    "Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; WOW64; Trident/5.0)",
-    "Mozilla/5.0 (Windows NT 6.1; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (Windows NT 6.2; WOW64; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (Windows NT 10.0; WOW64; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.0; Trident/5.0)",
-    "Mozilla/5.0 (Windows NT 6.3; WOW64; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; Trident/5.0)",
-    "Mozilla/5.0 (Windows NT 6.1; Win64; x64; Trident/7.0; rv:11.0) like Gecko",
-    "Mozilla/5.0 (compatible; MSIE 10.0; Windows NT 6.1; WOW64; Trident/6.0)",
-    "Mozilla/5.0 (compatible; MSIE 10.0; Windows NT 6.1; Trident/6.0)",
-    "Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 5.1; Trident/4.0; .NET CLR 2.0.50727; .NET CLR 3.0.4506.2152; .NET CLR 3.5.30729)"
-    # Safari
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.16 (KHTML, like Gecko) RockMelt/0.9.50.549 Chrome/10.0.648.205 Safari/534.16"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.18 (KHTML, like Gecko) Chrome/11.0.661.0 Safari/534.18"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.19 (KHTML, like Gecko) Chrome/11.0.661.0 Safari/534.19"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.21 (KHTML, like Gecko) Chrome/11.0.678.0 Safari/534.21"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.21 (KHTML, like Gecko) Chrome/11.0.682.0 Safari/534.21"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.3 (KHTML, like Gecko) Chrome/6.0.458.1 Safari/534.3"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.3 (KHTML, like Gecko) Chrome/6.0.461.0 Safari/534.3"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.3 (KHTML, like Gecko) Chrome/6.0.472.53 Safari/534.3"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.3 (KHTML, like Gecko) Iron/6.0.475 Safari/534"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.30 (KHTML, like Gecko) Chrome/12.0.724.100 Safari/534.30"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.6 (KHTML, like Gecko) Chrome/7.0.500.0 Safari/534.6"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Chrome/7.0.514.0 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Chrome/7.0.517.41 Safari/534.7 ChromePlus/1.5.0.0 ChromePlus/1.5.0.0"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Chrome/7.0.517.41 Safari/534.7 ChromePlus/1.5.0.0alpha1"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Flock/3.5.2.4599 Chrome/7.0.517.442 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Iron/7.0.520.0 Chrome/7.0.520.0 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Iron/7.0.520.1 Chrome/7.0.520.1 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) Iron/7.0.520.1 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) RockMelt/0.8.36.116 Chrome/7.0.517.44 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.7 (KHTML, like Gecko) RockMelt/0.8.36.128 Chrome/7.0.517.44 Safari/534.7"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/534.9 (KHTML, like Gecko) Chrome/7.0.531.0 Safari/534.9"
-    "Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/525.19 (KHTML, like Gecko) Iron/0.2.152.0 Safari/13657880.525",
-]
-
-
-def parse_tg_date(dd):
-    year = str(format(dd.year, "02d"))
-    month = str(format(dd.month, "02d"))
-    day = str(format(dd.day, "02d"))
-    hour = str(format(dd.hour, "02d"))
-    minute = str(format(dd.minute, "02d"))
-    second = str(format(dd.second, "02d"))
-    date = year + "-" + month + "-" + day
-    mtime = hour + ":" + minute + ":" + second
-    timestamp = date + "T" + mtime + "+00:00"
-    return {"timestamp":timestamp, "date":date, "mtime":mtime}
-
-
-def populate_user(user, group_or_chat):
-    if user.username:
-        username = user.username
-    else:
-        username = "n/a"
-    if user.first_name:
-        first_name = user.first_name
-    else:
-        first_name = ""
-    if user.last_name:
-        last_name = user.last_name
-    else:
-        last_name = ""
-    if user.phone:
-        phone = user.phone
-    else:
-        phone = "n/a"
-    full_name = (first_name + " " + last_name).strip()
-    return [username, full_name, user.id, phone, group_or_chat]
-
-
-def process_message(mess, user_lang):
-
-    if mess is not None:
-        mess_txt = '"' + mess + '"'
-    else:
-        mess_txt = "none"
-
-    if mess_txt != "none":
-        translator = Translator()
-        detection = translator.detect(mess_txt)
-        language_code = detection.lang
-        translation_confidence = detection.confidence
-        translation = translator.translate(mess_txt, dest=user_lang)
-        original_language = translation.src
-        translated_text = translation.text
-    else:
-        original_language = user_lang
-        translated_text = "n/a"
-        translation_confidence = "n/a"
-
-    return {
-        "original_language": original_language,
-        "translated_text": translated_text,
-        "translation_confidence": translation_confidence,
-        "message_text": mess_txt,
-    }
 
 
 @click.command()
@@ -198,24 +88,10 @@ def process_message(mess, user_lang):
     default=False,
     help="Enable replies analysis in channels.",
 )
-def cli(target, comprehensive, media, forwards, user, location, alt, json, export, replies):
-
-    print(
-        Fore.GREEN
-        + """
-      ______     __                 __  __
-     /_  __/__  / /__  ____  ____ _/ /_/ /_  __  __
-      / / / _ \/ / _ \/ __ \/ __ `/ __/ __ \/ / / /
-     / / /  __/ /  __/ /_/ / /_/ / /_/ / / / /_/ /
-    /_/  \___/_/\___/ .___/\__,_/\__/_/ /_/\__, /
-                   /_/                    /____/
-    -- An OSINT toolkit for investigating Telegram chats.
-    -- Developed by @jordanwildon | Version 2.1.10.
-    """
-    )
-
-    print(Style.RESET_ALL)
-
+def cli(
+    target, comprehensive, media, forwards, user, location, alt, json, export, replies
+):
+    print_banner()
     telepathy_file = "./telepathy_files/"
     try:
         os.makedirs(telepathy_file)
@@ -306,6 +182,10 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
             if character.isalnum():
                 alphanumeric += character
 
+        if "https://t.me/+" in t:
+            t = t.replace('https://t.me/+', 'https://t.me/joinchat/')
+
+
         save_directory = telepathy_file + alphanumeric
         try:
             os.makedirs(save_directory)
@@ -346,21 +226,10 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                     except FileExistsError:
                         pass
 
-
-
                 if basic == True:
-                    print(
-                        Fore.GREEN + " [!] " + Style.RESET_ALL + "Performing basic scan"
-                    )
-
+                    color_print_green(" [!] ", "Performing basic scan")
                 elif comp_check == True:
-                    print(
-                        Fore.GREEN
-                        + " [!] "
-                        + Style.RESET_ALL
-                        + "Performing comprehensive scan"
-                    )
-
+                    color_print_green(" [!] ", "Performing comprehensive scan")
                     file_archive = (
                         save_directory
                         + "/"
@@ -379,13 +248,7 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                     )
 
                 if forwards_check == True:
-                    print(
-                        Fore.GREEN
-                        + " [!] "
-                        + Style.RESET_ALL
-                        + "Forwards will be fetched"
-                    )
-
+                    color_print_green(" [!] ", "Forwards will be fetched")
                     file_forwards = (
                         save_directory
                         + "/edgelists/"
@@ -394,7 +257,6 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                         + filetime_clean
                         + "_edgelist.csv"
                     )
-
                     forward_directory = save_directory + "/edgelists/"
 
                     try:
@@ -408,16 +270,7 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
 
                 if basic is True or comp_check is True:
 
-                    print(
-                        "\n"
-                        + Fore.GREEN
-                        + " [-] "
-                        + Style.RESET_ALL
-                        + "Fetching details for "
-                        + t
-                        + "..."
-                    )
-
+                    color_print_green(" [-] ", "Fetching details for " + t + "...")
                     memberlist_directory = save_directory + "/memberlists"
 
                     try:
@@ -446,58 +299,29 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                         break
 
                     if entity.username:
+                        name = entity.title
                         group_url = "http://t.me/" + entity.username
                         group_username = entity.username
-                        s = requests.Session()
-                        s.max_redirects = 10
-                        s.headers["User-Agent"] = random.choice(user_agent)
-                        URL = s.get(group_url)
-                        URL.encoding = "utf-8"
-                        html_content = URL.text
-                        soup = BeautifulSoup(html_content, "html.parser")
-                        name = entity.title
-                        try:
-                            group_description = soup.find(
-                                "div", {"class": ["tgme_page_description"]}
-                            ).text
-                            descript = Fore.GREEN + "Description: " + Style.RESET_ALL
-                            prefix = descript + " "
-                            preferredWidth = 70
-                            wrapper_d = textwrap.TextWrapper(
-                                initial_indent=descript,
-                                width=preferredWidth,
-                                subsequent_indent="                  ",
-                            )
-                        except:
-                            group_description = "None"
-                            descript = Fore.GREEN + "Description: " + Style.RESET_ALL
-                            prefix = descript + " "
-                            preferredWidth = 70
-                            wrapper_d = textwrap.TextWrapper(
-                                initial_indent=descript,
-                                width=preferredWidth,
-                                subsequent_indent="                  ",
-                            )
-                        try:
-                            group_participants = soup.find(
-                                "div", {"class": ["tgme_page_extra"]}
-                            ).text
-                            sep = "members"
-                            stripped = group_participants.split(sep, 1)[0]
-                            total_participants = (
-                                stripped.replace(" ", "")
-                                .replace("members", "")
-                                .replace("subscribers", "")
-                                .replace("member", "")
-                            )
-                        except:
-                            total_participants = "Not found"  # could be due to restriction, might need to mention
-
+                        web_req = parse_html_page(group_url)
+                    elif "https://t.me/" in t:
+                        group_url = t
+                        web_req = parse_html_page(group_url)
+                        group_username = "Private group"
                     else:
                         group_url = "Private group"
                         group_username = "Private group"
-                        total_participants = "Not found"
-                        group_description = "Not found"
+
+
+                    group_description = web_req["group_description"]
+                    total_participants = web_req["total_participants"]
+                    preferredWidth = 70
+                    descript = Fore.GREEN + "Description: " + Style.RESET_ALL + web_req["group_description"]
+                    prefix = descript + " "
+                    wrapper_d = textwrap.TextWrapper(
+                        initial_indent=descript,
+                        width=preferredWidth,
+                        subsequent_indent="                  ",
+                    )
 
                     if entity.broadcast is True:
                         chat_type = "Channel"
@@ -534,6 +358,7 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                         all_participants = []
                         all_participants = await client.get_participants(t, limit=5000)
 
+                        members_df = None
                         for user in all_participants:
                             members_df = pd.DataFrame(
                                 members,
@@ -547,21 +372,22 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                             )
                             members.append(populate_user(user, t))
 
-                        with open(
-                            memberlist_filename, "w+", encoding="utf-8"
-                        ) as save_members:
-                            members_df.to_csv(save_members, sep=";")
+                        if members_df is not None:
+                            with open(
+                                memberlist_filename, "w+", encoding="utf-8"
+                            ) as save_members:
+                                members_df.to_csv(save_members, sep=";")
 
-                        if json_check == True:
-                            members_df.to_json(
-                                json_file + alphanumeric + "_memberlist.json",
-                                orient="records",
-                                compression="infer",
-                                lines=True,
-                                index=True,
-                            )
-                        else:
-                            pass
+                            if json_check == True:
+                                members_df.to_json(
+                                    json_file + alphanumeric + "_memberlist.json",
+                                    orient="records",
+                                    compression="infer",
+                                    lines=True,
+                                    index=True,
+                                )
+                            else:
+                                pass
 
                         found_participants = len(all_participants)
                         found_participants = int(found_participants)
@@ -572,104 +398,52 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                     log = []
 
                     if chat_type != "Channel":
-                        print(
-                            "\n"
-                            + Fore.GREEN
-                            + " [+] Memberlist fetched"
-                            + Style.RESET_ALL
-                        )
+                        print("\n")
+                        color_print_green(" [+] Memberlist fetched", "")
                     else:
                         pass
-                    print(Fore.GREEN + "  ┬  Chat details" + Style.RESET_ALL)
 
-                    print(
-                        Fore.GREEN
-                        + "  ├  Title: "
-                        + Style.RESET_ALL
-                        + str(entity.title)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  "
-                        + Style.RESET_ALL
-                        + wrapper_d.fill(group_description)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Total participants: "
-                        + Style.RESET_ALL
-                        + str(total_participants)
+                    color_print_green("  ┬  Chat details", "")
+                    color_print_green("  ├  Title: ", str(entity.title))
+                    color_print_green("  ├  ", wrapper_d.fill(group_description))
+                    color_print_green(
+                        "  ├  Total participants: ", str(total_participants)
                     )
 
                     if chat_type != "Channel":
-                        print(
-                            Fore.GREEN
-                            + "  ├  Participants found: "
-                            + Style.RESET_ALL
-                            + str(found_participants)
+                        color_print_green(
+                            "  ├  Participants found: ",
+                            str(found_participants)
                             + " ("
                             + str(format(found_percentage, ".2f"))
-                            + "%)"
+                            + "%)",
                         )
                     else:
                         found_participants = "N/A"
 
-                    print(
-                        Fore.GREEN
-                        + "  ├  Username: "
-                        + Style.RESET_ALL
-                        + str(group_username)
-                    )
-                    print(Fore.GREEN + "  ├  URL: " + Style.RESET_ALL + str(group_url))
-                    print(
-                        Fore.GREEN
-                        + "  ├  Chat type: "
-                        + Style.RESET_ALL
-                        + str(chat_type)
-                    )
-                    print(
-                        Fore.GREEN + "  ├  Chat id: " + Style.RESET_ALL + str(entity.id)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Access hash: "
-                        + Style.RESET_ALL
-                        + str(entity.access_hash)
-                    )
+                    color_print_green("  ├  Username: ", str(group_username))
+                    color_print_green("  ├  URL: ", str(group_url))
+                    color_print_green("  ├  Chat type: ", str(chat_type))
+                    color_print_green("  ├  Chat id: ", str(entity.id))
+                    color_print_green("  ├  Access hash: ", str(entity.access_hash))
 
                     if chat_type == "Channel":
                         scam_status = str(entity.scam)
-                        print(
-                            Fore.GREEN
-                            + "  ├  Scam: "
-                            + Style.RESET_ALL
-                            + str(scam_status)
-                        )
+                        color_print_green("  ├  Scam: ", str(scam_status))
                     else:
                         scam_status = "N/A"
 
-                    print(
-                        Fore.GREEN
-                        + "  ├  First post date: "
-                        + Style.RESET_ALL
-                        + str(first_post)
-                    )
+                    color_print_green("  ├  First post date: ", str(first_post))
 
                     if chat_type != "Channel":
-                        print(
-                            Fore.GREEN
-                            + "  ├  Memberlist saved to: "
-                            + Style.RESET_ALL
-                            + memberlist_filename
-                        )  # tidy in display
+                        color_print_green(
+                            "  ├  Memberlist saved to: ", memberlist_filename
+                        )
 
-                    print(
-                        Fore.GREEN
-                        + "  └  "
-                        + Style.RESET_ALL
-                        + wrapper_r.fill(group_status)
-                        + "\n"
+                    color_print_green(
+                        "  └  ", wrapper_r.fill(group_status)
                     )
+                    print("\n")
 
                     log.append(
                         [
@@ -718,13 +492,9 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                         )
 
                     if forwards_check is True and comp_check is False:
-                        print(
-                            Fore.GREEN
-                            + " [-] "
-                            + Style.RESET_ALL
-                            + "Calculating number of forwarded messages..."
+                        color_print_green(
+                            " [-] ", "Calculating number of forwarded messages..."
                         )
-
                         forwards_list = []
                         forward_count = 0
                         private_count = 0
@@ -747,13 +517,8 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                             if message.forward is not None:
                                 forward_count += 1
 
-                        print(
-                            "\n"
-                            + Fore.GREEN
-                            + " [-] "
-                            + Style.RESET_ALL
-                            + "Fetching forwarded messages..."
-                        )
+                        print("\n")
+                        color_print_green(" [-] ", "Fetching forwarded messages...")
 
                         progress_bar = (
                             Fore.GREEN + " [-] " + Style.RESET_ALL + "Progress: "
@@ -770,7 +535,9 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                         if f_from_id is not None:
                                             ent = await client.get_entity(f_from_id)
                                             username = ent.username
-                                            timestamp = parse_tg_date(message.date)["timestamp"]
+                                            timestamp = parse_tg_date(message.date)[
+                                                "timestamp"
+                                            ]
 
                                             substring = "PeerUser"
                                             string = str(f_from_id)
@@ -888,68 +655,37 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                             df02 = forwards_df.From.unique()
                             unique_forwards = len(df02)
 
-                            print(
-                                "\n"
-                                + Fore.GREEN
-                                + " [+] Forward scrape complete"
-                                + Style.RESET_ALL
+                            print("\n")
+                            color_print_green(" [+] Forward scrape complete", "")
+                            color_print_green("  ┬  Statistics", "")
+                            color_print_green(
+                                "  ├  Forwarded messages found: ", str(forward_count)
                             )
-                            print(Fore.GREEN + "  ┬  Statistics" + Style.RESET_ALL)
-                            print(
-                                Fore.GREEN
-                                + "  ├  Forwarded messages found: "
-                                + Style.RESET_ALL
-                                + str(forward_count)
+                            color_print_green(
+                                "  ├  Forwards from active public chats: ",
+                                str(forwards_found),
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Forwards from active public chats: "
-                                + Style.RESET_ALL
-                                + str(forwards_found)
+                            color_print_green(
+                                "  ├  Unique forward sources: ", str(unique_forwards)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Unique forward sources: "
-                                + Style.RESET_ALL
-                                + str(unique_forwards)
+                            color_print_green(
+                                "  ├  Top forward source 1: ", str(forward_one)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Top forward source 1: "
-                                + Style.RESET_ALL
-                                + str(forward_one)
+                            color_print_green(
+                                "  ├  Top forward source 2: ", str(forward_two)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Top forward source 2: "
-                                + Style.RESET_ALL
-                                + str(forward_two)
+                            color_print_green(
+                                "  ├  Top forward source 3: ", str(forward_three)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Top forward source 3: "
-                                + Style.RESET_ALL
-                                + str(forward_three)
+                            color_print_green(
+                                "  ├  Top forward source 4: ", str(forward_four)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Top forward source 4: "
-                                + Style.RESET_ALL
-                                + str(forward_four)
+                            color_print_green(
+                                "  ├  Top forward source 5: ", str(forward_five)
                             )
-                            print(
-                                Fore.GREEN
-                                + "  ├  Top forward source 5: "
-                                + Style.RESET_ALL
-                                + str(forward_five)
-                            )
-                            print(
-                                Fore.GREEN
-                                + "  └  Edgelist saved to: "
-                                + Style.RESET_ALL
-                                + edgelist_file
-                                + "\n"
-                            )
+                            color_print_green("  └  Edgelist saved to: ", edgelist_file)
+                            print("\n")
+
                         else:
                             print(
                                 "\n"
@@ -978,19 +714,13 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                             private_count = 0
 
                             if media_archive is True:
-                                print(
-                                    Fore.GREEN
-                                    + " [!] "
-                                    + Style.RESET_ALL
-                                    + "Media content will be archived"
+                                color_print_green(
+                                    " [!] ", "Media content will be archived"
                                 )
 
-                            print(
-                                Fore.GREEN
-                                + "\n"
-                                + " [!] "
-                                + Style.RESET_ALL
-                                + "Calculating number of messages..."
+                            print("\n")
+                            color_print_green(
+                                " [!] ", "Calculating number of messages..."
                             )
 
                             message_count = 0
@@ -999,14 +729,8 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                 if message is not None:
                                     message_count += 1
 
-                            print(
-                                Fore.GREEN
-                                + "\n"
-                                + " [-] "
-                                + Style.RESET_ALL
-                                + "Fetching message archive..."
-                            )
-
+                            print("\n")
+                            color_print_green(" [-] ", "Fetching message archive...")
                             progress_bar = (
                                 Fore.GREEN + " [-] " + Style.RESET_ALL + "Progress: "
                             )
@@ -1024,6 +748,7 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                     t, limit=None
                                 ):
                                     if message is not None:
+
                                         try:
 
                                             c_archive = pd.DataFrame(
@@ -1068,7 +793,11 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                                         ],
                                                     )
 
-                                            if message.replies and reply_analysis and chat_type == "Channel" :
+                                            if (
+                                                message.replies
+                                                and reply_analysis
+                                                and chat_type == "Channel"
+                                            ):
                                                 if message.replies.replies > 0:
                                                     c_repliers = pd.DataFrame(
                                                         user_replier_list,
@@ -1151,7 +880,9 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                             else:
                                                 nameID = to_ent.id
 
-                                            timestamp = parse_tg_date(message.date)["timestamp"]
+                                            timestamp = parse_tg_date(message.date)[
+                                                "timestamp"
+                                            ]
                                             reply = message.reply_to_msg_id
 
                                             _mess = process_message(
@@ -1441,70 +1172,44 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                 # one day this'll work out sleeping times
                                 # print(c_t_stats)
 
-                            print(
-                                "\n"
-                                + Fore.GREEN
-                                + " [+] Chat archive saved"
-                                + Style.RESET_ALL
+                            print("\n")
+                            color_print_green(" [+] Chat archive saved", "")
+                            color_print_green("  ┬  Chat statistics", "")
+                            color_print_green(
+                                "  ├  Number of messages found: ", str(messages_found)
                             )
-                            print(Fore.GREEN + "  ┬  Chat statistics" + Style.RESET_ALL)
-                            print(
-                                Fore.GREEN
-                                + "  ├  Number of messages found: "
-                                + Style.RESET_ALL
-                                + str(messages_found)
-                            )
+
                             if chat_type != "Channel":
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Top poster 1: "
-                                    + Style.RESET_ALL
-                                    + str(poster_one)
+                                color_print_green(
+                                    "  ├  Top poster 1: ", str(poster_one)
                                 )
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Top poster 2: "
-                                    + Style.RESET_ALL
-                                    + str(poster_two)
+                                color_print_green(
+                                    "  ├  Top poster 2: ", str(poster_two)
                                 )
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Top poster 3: "
-                                    + Style.RESET_ALL
-                                    + str(poster_three)
+                                color_print_green(
+                                    "  ├  Top poster 3: ", str(poster_three)
                                 )
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Top poster 4: "
-                                    + Style.RESET_ALL
-                                    + str(poster_four)
+                                color_print_green(
+                                    "  ├  Top poster 4: ", str(poster_four)
                                 )
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Top poster 5: "
-                                    + Style.RESET_ALL
-                                    + str(poster_five)
+                                color_print_green(
+                                    "  ├  Top poster 5: ", str(poster_five)
                                 )
-                                print(
-                                    Fore.GREEN
-                                    + "  ├  Total unique posters: "
-                                    + Style.RESET_ALL
-                                    + str(unique_active)
+                                color_print_green(
+                                    "  ├  Total unique posters: ", str(unique_active)
                                 )
+
                                 # add a figure for unique current posters who are active
                             else:
                                 pass
                                 # timestamp analysis
-                            #    print(Fore.GREEN
-                            #          + '  ├  Number of messages: '
-                            #          + Style.RESET_ALL
-                            #          + str(message_count))
+                                #    print(Fore.GREEN
+                                #            + '  ├  Number of messages: '
+                                #          + Style.RESET_ALL
+                                #          + str(message_count))
 
-                            print(
-                                Fore.GREEN
-                                + "  └  Archive saved to: "
-                                + Style.RESET_ALL
-                                + str(file_archive)
+                            color_print_green(
+                                "  └  Archive saved to: ", str(file_archive)
                             )
 
                             if len(replies_list) > 0:
@@ -1512,31 +1217,17 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                 if user_replier_list == 0:
                                     middle_char = "└"
 
-                                print(
-                                    "\n"
-                                    + Fore.GREEN
-                                    + " [+] Replies analysis "
-                                    + Style.RESET_ALL
-                                )
-                                print(
-                                    Fore.GREEN
-                                    + "  ┬  Chat statistics"
-                                    + Style.RESET_ALL
-                                )
-                                print(
-                                    Fore.GREEN
-                                    + "  "
-                                    + middle_char
-                                    + "  Archive of replies saved to: "
-                                    + Style.RESET_ALL
-                                    + str(reply_file_archive)
+                                print("\n")
+                                color_print_green(" [+] Replies analysis ", "")
+                                color_print_green("  ┬  Chat statistics", "")
+                                color_print_green(
+                                    f"  {middle_char}  Archive of replies saved to: ",
+                                    str(reply_file_archive),
                                 )
                                 if len(user_replier_list) > 0:
-                                    print(
-                                        Fore.GREEN
-                                        + "  └  Active members list who replied to messages, saved to: "
-                                        + Style.RESET_ALL
-                                        + str(reply_memberlist_filename)
+                                    color_print_green(
+                                        "  └  Active members list who replied to messages, saved to: ",
+                                        str(reply_memberlist_filename),
                                     )
 
                             if forwards_check is True:
@@ -1598,92 +1289,61 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                                     c_f_unique = c_forwards.From.unique()
                                     unique_forwards = len(c_f_unique)
 
-                                    print(
-                                        "\n"
-                                        + Fore.GREEN
-                                        + " [+] Edgelist saved"
-                                        + Style.RESET_ALL
+                                    print("\n")
+                                    color_print_green(" [+] Edgelist saved", "")
+                                    color_print_green(
+                                        "  ┬  Forwarded message statistics", ""
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ┬  Forwarded message statistics"
-                                        + Style.RESET_ALL
+                                    color_print_green(
+                                        "  ├  Forwarded messages found: ",
+                                        str(forward_count),
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Forwarded messages found: "
-                                        + Style.RESET_ALL
-                                        + str(forward_count)
+                                    color_print_green(
+                                        "  ├  Forwards from active public chats: ",
+                                        str(forwards_found),
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Forwards from active public chats: "
-                                        + Style.RESET_ALL
-                                        + str(forwards_found)
+                                    color_print_green(
+                                        "  ├  Forwards from private (or now private) chats: ",
+                                        str(private_count),
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Forwards from private (or now private) chats: "
-                                        + Style.RESET_ALL
-                                        + str(private_count)
+                                    color_print_green(
+                                        "  ├  Unique forward sources: ",
+                                        str(unique_forwards),
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Unique forward sources: "
-                                        + Style.RESET_ALL
-                                        + str(unique_forwards)
+                                    color_print_green(
+                                        "  ├  Top forward source 1: ", str(forward_one)
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Top forward source 1: "
-                                        + Style.RESET_ALL
-                                        + str(forward_one)
+                                    color_print_green(
+                                        "  ├  Top forward source 2: ", str(forward_two)
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Top forward source 2: "
-                                        + Style.RESET_ALL
-                                        + str(forward_two)
+                                    color_print_green(
+                                        "  ├  Top forward source 3: ",
+                                        str(forward_three),
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Top forward source 3: "
-                                        + Style.RESET_ALL
-                                        + str(forward_three)
+                                    color_print_green(
+                                        "  ├  Top forward source 4: ", str(forward_four)
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Top forward source 4: "
-                                        + Style.RESET_ALL
-                                        + str(forward_four)
+                                    color_print_green(
+                                        "  ├  Top forward source 5: ", str(forward_five)
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  ├  Top forward source 5: "
-                                        + Style.RESET_ALL
-                                        + str(forward_five)
+                                    color_print_green(
+                                        "  └  Edgelist saved to: ", edgelist_file
                                     )
-                                    print(
-                                        Fore.GREEN
-                                        + "  └  Edgelist saved to: "
-                                        + Style.RESET_ALL
-                                        + file_forwards
-                                        + "\n"
-                                    )  # make edgelist gephi compatible
-                                else:
-                                    print(
-                                        "\n"
-                                        + Fore.GREEN
-                                        + " [!] Insufficient forwarded messages found"
-                                        + Style.RESET_ALL
-                                    )
+                                    print("\n")
 
+                                else:
+                                    print("\n")
+                                    color_print_green(
+                                        " [!] Insufficient forwarded messages found",
+                                        edgelist_file,
+                                    )
                             else:
                                 pass
 
                 if user_check == True:
                     my_user = None
                     try:
+
                         print(
                             Fore.GREEN
                             + " [+] "
@@ -1722,70 +1382,19 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                         else:
                             user_restrictions = "None"
 
-                        print(
-                            Fore.GREEN
-                            + "  ├  Username: "
-                            + Style.RESET_ALL
-                            + str(my_user.username)
+                        color_print_green(" [+] ", "User details for " + t)
+                        color_print_green("  ├  Username: ", str(my_user.username))
+                        color_print_green("  ├  Name: ", str(user_full_name))
+                        color_print_green("  ├  Verification: ", str(my_user.verified))
+                        color_print_green("  ├  Photo ID: ", str(user_photo))
+                        color_print_green("  ├  Phone number: ", str(my_user.phone))
+                        color_print_green(
+                            "  ├  Access hash: ", str(my_user.access_hash)
                         )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Name: "
-                            + Style.RESET_ALL
-                            + str(user_full_name)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Verification: "
-                            + Style.RESET_ALL
-                            + str(my_user.verified)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Photo ID: "
-                            + Style.RESET_ALL
-                            + str(user_photo)
-                        )
-                        # print(Fore.GREEN
-                        #      + '  ├  Status: '
-                        #      + Style.RESET_ALL
-                        #      + str(my_user.status))
-                        print(
-                            Fore.GREEN
-                            + "  ├  Phone number: "
-                            + Style.RESET_ALL
-                            + str(my_user.phone)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Access hash: "
-                            + Style.RESET_ALL
-                            + str(my_user.access_hash)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Language: "
-                            + Style.RESET_ALL
-                            + str(my_user.lang_code)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Bot: "
-                            + Style.RESET_ALL
-                            + str(my_user.bot)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  ├  Scam: "
-                            + Style.RESET_ALL
-                            + str(my_user.scam)
-                        )
-                        print(
-                            Fore.GREEN
-                            + "  └  Restrictions: "
-                            + Style.RESET_ALL
-                            + str(user_restrictions)
-                        )
+                        color_print_green("  ├  Language: ", str(my_user.lang_code))
+                        color_print_green("  ├  Bot: ", str(my_user.bot))
+                        color_print_green("  ├  Scam: ", str(my_user.scam))
+                        color_print_green("  └  Restrictions: ", str(user_restrictions))
 
                     except ValueError:
                         pass
@@ -1888,46 +1497,15 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
 
                     total = len(locations_list)
 
-                    print(Fore.GREEN + " [+] Users located" + Style.RESET_ALL)
-                    print(
-                        Fore.GREEN
-                        + "  ├  Users within 500m:  "
-                        + Style.RESET_ALL
-                        + str(d_500)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Users within 1000m: "
-                        + Style.RESET_ALL
-                        + str(d_1000)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Users within 2000m: "
-                        + Style.RESET_ALL
-                        + str(d_2000)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Users within 3000m: "
-                        + Style.RESET_ALL
-                        + str(d_3000)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  ├  Total users found:  "
-                        + Style.RESET_ALL
-                        + str(total)
-                    )
-                    print(
-                        Fore.GREEN
-                        + "  └  Location list saved to: "
-                        + Style.RESET_ALL
-                        + save_file
-                    )
+                    color_print_green(" [+] Users located", "")
+                    color_print_green("  ├  Users within 500m:  ", str(d_500))
+                    color_print_green("  ├  Users within 1000m: ", str(d_1000))
+                    color_print_green("  ├  Users within 2000m: ", str(d_2000))
+                    color_print_green("  ├  Users within 3000m: ", str(d_3000))
+                    color_print_green("  ├  Total users found:  ", str(total))
+                    color_print_green("  └  Location list saved to: ", save_file)
 
                     # can also do the same for channels with similar output file to users
-
                     # may one day add trilateration to find users closest to exact point
 
                 if export == True:
@@ -1941,54 +1519,10 @@ def cli(target, comprehensive, media, forwards, user, location, alt, json, expor
                             if Dialog.entity.username:
                                 group_url = "http://t.me/" + Dialog.entity.username
                                 group_username = Dialog.entity.username
-                                s = requests.Session()
-                                s.max_redirects = 10
-                                s.headers["User-Agent"] = random.choice(user_agent)
-                                URL = s.get(group_url)
-                                URL.encoding = "utf-8"
-                                html_content = URL.text
-                                soup = BeautifulSoup(html_content, "html.parser")
-                                name = Dialog.entity.title
-                                try:
-                                    group_description = soup.find(
-                                        "div", {"class": ["tgme_page_description"]}
-                                    ).text
-                                    descript = (
-                                        Fore.GREEN + "Description: " + Style.RESET_ALL
-                                    )
-                                    prefix = descript + " "
-                                    preferredWidth = 70
-                                    wrapper_d = textwrap.TextWrapper(
-                                        initial_indent=descript,
-                                        width=preferredWidth,
-                                        subsequent_indent="                  ",
-                                    )
-                                except:
-                                    group_description = "None"
-                                    descript = (
-                                        Fore.GREEN + "Description: " + Style.RESET_ALL
-                                    )
-                                    prefix = descript + " "
-                                    preferredWidth = 70
-                                    wrapper_d = textwrap.TextWrapper(
-                                        initial_indent=descript,
-                                        width=preferredWidth,
-                                        subsequent_indent="                  ",
-                                    )
-                                try:
-                                    group_participants = soup.find(
-                                        "div", {"class": ["tgme_page_extra"]}
-                                    ).text
-                                    sep = "members"
-                                    stripped = group_participants.split(sep, 1)[0]
-                                    total_participants = (
-                                        stripped.replace(" ", "")
-                                        .replace("members", "")
-                                        .replace("subscribers", "")
-                                        .replace("member", "")
-                                    )
-                                except:
-                                    total_participants = "Not found"  # could be due to restriction, might need to mention
+
+                                web_req = parse_html_page(group_url)
+                                group_description = web_req["group_description"]
+                                total_participants = web_req["total_participants"]
 
                                 if Dialog.entity.broadcast is True:
                                     chat_type = "Channel"
